disable 32-bit build; fix test step

+++ src/ast.zig
@@ -637,7 +637,7 @@ pub const Payload = struct {
 
         pub const ArrayTypeInfo = struct {
             elem_type: Node,
-            len: usize,
+            len: u64,
         };
     };
